news_reader/requirements.txt
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
courlan==1.3.2
cssselect==1.3.0
dateparser==1.2.2
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
htmldate==1.9.3
httplib2==0.31.0
idna==3.11
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jusText==3.0.2
lxml==5.4.0
lxml_html_clean==0.4.3
macholib==1.16.3
MarkupSafe==3.0.3
narwhals==2.10.1
newspaper3k==0.2.8
nltk==3.9.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pillow==12.0.0
playwright==1.55.0
playwright-stealth==2.0.0
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
pyee==13.0.0
pyinstaller-hooks-contrib==2025.9
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-file==3.0.1
rpds-py==0.28.0
rsa==4.9.1
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8
streamlit==1.51.0
tenacity==9.1.2
tinysegmenter==0.3
tld==0.13.1
tldextract==5.3.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
